+++ apps/api-harmonization/src/blocks/article-list/article-list.mapper.ts
@@ -7,7 +7,6 @@ import { ArticleListBlock } from './article-list.model';
 export const mapArticleList = (
     cms: CMS.Model.ArticleListBlock.ArticleListBlock,
     articles: Articles.Model.Articles,
-
     locale: string,
 ): ArticleListBlock => {
     return {
@@ -15,6 +14,13 @@ export const mapArticleList = (
         id: cms.id,
         title: cms.title,
         description: cms.description,
+        categoryLink:
+            cms.categorySlug && cms.parent?.slug
+                ? {
+                      url: `/${cms.parent?.slug}/${cms.categorySlug}`,
+                      label: cms.labels.seeAllArticles,
+                  }
+                : undefined,
         items: {
             ...articles,
             data: articles.data.map((article) => mapArticle(article, cms, locale)),