+++ apps/api-harmonization/src/blocks/order-details/order-details.module.ts
@@ -1,8 +1,9 @@
+import { HttpModule } from '@nestjs/axios';
 import { DynamicModule, Module } from '@nestjs/common';
 
 import { ApiConfig } from '@o2s/framework/modules';
 
-import { CMS, Orders, Products } from '../../models';
+import { CMS, Orders } from '../../models';
 
 import { OrderDetailsController } from './order-details.controller';
 import { OrderDetailsService } from './order-details.service';
@@ -12,9 +13,10 @@ export class OrderDetailsBlockModule {
     static register(_config: ApiConfig): DynamicModule {
         return {
             module: OrderDetailsBlockModule,
-            providers: [OrderDetailsService, CMS.Service, Orders.Service, Products.Service],
+            providers: [OrderDetailsService, CMS.Service, Orders.Service],
             controllers: [OrderDetailsController],
             exports: [OrderDetailsService],
+            imports: [HttpModule],
         };
     }
 }