+++ apps/api-harmonization/src/modules/page/page.model.ts
@@ -3,6 +3,7 @@ import { CMS, Models } from '@o2s/framework/modules';
 import {
     Article,
     ArticleList,
+    ArticleSearch,
     Category,
     CategoryList,
     Faq,
@@ -73,7 +74,7 @@ export type Blocks =
     | ArticleList.Model.ArticleListBlock['__typename']
     | Category.Model.CategoryBlock['__typename']
     | Article.Model.ArticleBlock['__typename']
-    // BLOCK REGISTER
+    | ArticleSearch.Model.ArticleSearchBlock['__typename']
     | TicketList.Model.TicketListBlock['__typename']
     | TicketDetails.Model.TicketDetailsBlock['__typename']
     | NotificationList.Model.NotificationListBlock['__typename']
@@ -92,3 +93,4 @@ export type Blocks =
     | OrderDetails.Model.OrderDetailsBlock['__typename']
     | QuickLinks.Model.QuickLinksBlock['__typename']
     | CategoryList.Model.CategoryListBlock['__typename'];
+// BLOCK REGISTER