+++ apps/frontend/src/components/Cards/ProductCard/ProductCard.types.ts
@@ -14,7 +14,7 @@ export interface ProductCardProps {
         label: string;
         url: string;
     };
-    image: Models.Media.Media;
+    image?: Models.Media.Media;
 }
 
 export interface Badge {