+++ packages/framework/src/modules/cms/cms.model.ts
@@ -34,3 +34,4 @@ export * as ArticleListBlock from './models/blocks/article-list.model';
 export * as CategoryBlock from './models/blocks/category.model';
 export * as CategoryListBlock from './models/blocks/category-list.model';
 export * as QuickLinksBlock from './models/blocks/quick-links.model';
+export * as ArticleSearchBlock from './models/blocks/article-search.model';