+++ packages/integrations/algolia/src/modules/search/models/article.model.ts
@@ -1,5 +1,18 @@
-import { Articles } from '@o2s/framework/modules';
-
-export class SearchEngineArticleModel extends Articles.Model.Article {
-    objectID!: string;
-}
+/* eslint-disable @typescript-eslint/no-explicit-any */
+export type SearchEngineArticleModel = {
+    id: string;
+    documentId: string;
+    slug: string;
+    locale?: string;
+    createdAt?: any;
+    updatedAt?: any;
+    publishedAt?: any;
+    SEO: {
+        title: string;
+        noIndex: boolean;
+        noFollow: boolean;
+        description: string;
+        keywords?: Array<{ keyword: string }>;
+        image?: { url: string; alternativeText?: string; width?: number; height?: number; name: string };
+    };
+};