+++ packages/integrations/mocked/src/modules/cms/mappers/blocks/cms.article-list.mapper.ts
@@ -5,13 +5,15 @@ const MOCK_ARTICLE_LIST_BLOCK_EN: CMS.Model.ArticleListBlock.ArticleListBlock = 
     title: 'Explore How-To Guides',
     description:
         'Discover step-by-step tutorials and helpful guides to get the most out of your products and solve common issues.',
-    categoryId: 'warranty-and-repair',
+    categorySlug: 'warranty-and-repair',
+    articlesToShow: 4,
     parent: {
         slug: '/help-and-support',
     },
     labels: {
         today: 'Today',
         yesterday: 'Yesterday',
+        seeAllArticles: 'See all articles',
     },
 };
 
@@ -20,13 +22,15 @@ const MOCK_ARTICLE_LIST_BLOCK_DE: CMS.Model.ArticleListBlock.ArticleListBlock = 
     title: 'Entdecke Anleitungen',
     description:
         'Entdecken Sie Schritt-für-Schritt-Anleitungen und hilfreiche Ratgeber, um das Beste aus Ihren Produkten herauszuholen und häufige Probleme zu lösen.',
-    categoryId: 'warranty-and-repair',
+    categorySlug: 'warranty-and-repair',
+    articlesToShow: 4,
     parent: {
         slug: '/hilfe-und-support',
     },
     labels: {
         today: 'Heute',
         yesterday: 'Gestern',
+        seeAllArticles: 'Alle Artikel anzeigen',
     },
 };
 
@@ -35,13 +39,15 @@ const MOCK_ARTICLE_LIST_BLOCK_PL: CMS.Model.ArticleListBlock.ArticleListBlock = 
     title: 'Przeglądaj poradniki',
     description:
         'Odkryj szczegółowe instrukcje i pomocne poradniki, aby w pełni wykorzystać możliwości swoich produktów i rozwiązać typowe problemy.',
-    categoryId: 'warranty-and-repair',
+    categorySlug: 'warranty-and-repair',
+    articlesToShow: 4,
     parent: {
         slug: '/pomoc-i-wsparcie',
     },
     labels: {
         today: 'Dzisiaj',
         yesterday: 'Wczoraj',
+        seeAllArticles: 'Zobacz wszystkie artykuły',
     },
 };