+++ packages/integrations/mocked/src/modules/resources/resources.service.ts
@@ -1,14 +1,25 @@
 import { Injectable } from '@nestjs/common';
 import { Observable, of } from 'rxjs';
 
-import { Resources } from '@o2s/framework/modules';
+import { Products, Resources } from '@o2s/framework/modules';
 
 import { responseDelay } from '@/utils/delay';
 
-import { mapAsset, mapAssets, mapService, mapServices } from './resources.mapper';
+import {
+    mapAsset,
+    mapAssets,
+    mapCompatibleServices,
+    mapFeaturedServices,
+    mapService,
+    mapServices,
+} from './resources.mapper';
 
 @Injectable()
 export class ResourcesService implements Resources.Service {
+    purchaseOrActivateService(_params: Resources.Request.GetServiceParams): Observable<void> {
+        throw new Error('Method not implemented.');
+    }
+
     purchaseOrActivateResource(_params: Resources.Request.GetResourceParams): Observable<void> {
         throw new Error('Method not implemented');
     }
@@ -30,5 +41,13 @@ export class ResourcesService implements Resources.Service {
 
     getAsset(params: Resources.Request.GetAssetParams): Observable<Resources.Model.Asset> {
         return of(mapAsset(params.id)).pipe(responseDelay());
+    }
+
+    getCompatibleServiceList(params: Resources.Request.GetAssetParams): Observable<Products.Model.Products> {
+        return of(mapCompatibleServices(params)).pipe(responseDelay());
+    }
+
+    getFeaturedServiceList(): Observable<Products.Model.Products> {
+        return of(mapFeaturedServices()).pipe(responseDelay());
     }
 }