+++ packages/create-o2s-app/CHANGELOG.md
@@ -1,5 +1,11 @@
 # create-o2s-app
 
+## 1.0.0
+
+### Major Changes
+
+- 0e0c816: Official stable release
+
 ## 0.13.0
 
 ### Minor Changes